+++ ultralytics/yolo/v8/classify/train.py
@@ -4,8 +4,8 @@ import torch
 from ultralytics.yolo import v8
 from ultralytics.yolo.data import build_classification_dataloader
 from ultralytics.yolo.engine.trainer import DEFAULT_CONFIG, BaseTrainer
-from ultralytics.yolo.nn import get_model
-from ultralytics.yolo.nn.tasks import ClassificationModel
+from ultralytics.nn import get_model
+from ultralytics.nn.tasks import ClassificationModel
 
 
 class ClassificationTrainer(BaseTrainer):
