+++ ultralytics/yolo/v8/detect/train.py
@@ -6,7 +6,7 @@ from ultralytics.yolo import v8
 from ultralytics.yolo.data import build_dataloader
 from ultralytics.yolo.data.dataloaders.v5loader import create_dataloader
 from ultralytics.yolo.engine.trainer import DEFAULT_CONFIG, BaseTrainer
-from ultralytics.yolo.nn.tasks import DetectionModel
+from ultralytics.nn.tasks import DetectionModel
 from ultralytics.yolo.utils import colorstr
 from ultralytics.yolo.utils.loss import BboxLoss
 from ultralytics.yolo.utils.metrics import smooth_BCE
