+++ ultralytics/yolo/v8/segment/train.py
@@ -5,7 +5,7 @@ import torch.nn.functional as F
 
 from ultralytics.yolo import v8
 from ultralytics.yolo.engine.trainer import DEFAULT_CONFIG
-from ultralytics.yolo.nn.tasks import SegmentationModel
+from ultralytics.nn.tasks import SegmentationModel
 from ultralytics.yolo.utils.metrics import FocalLoss, bbox_iou, smooth_BCE
 from ultralytics.yolo.utils.ops import crop_mask, xywh2xyxy
 from ultralytics.yolo.utils.plotting import plot_images, plot_results
